Upstream and downstream shouldn't have to go through authentication.
* Upstream and downstream don't have to go through authentication
There are two code paths for wse:
1) tcp | http | httpxe | wse
2) tcp | http | wse

In the second case, http layer doesn't get REALM_NAME security option for
upstream/downstream addresses. But in the first case, it does get REALM_NAME
security option. Hence, the security filters try to do authentication for
upstream and downstream requests. This needs to be fixed similar to the second case.

* cross-site-constraint is not getting propagated to httpxe layer while
constructing upstream and downstream resource addresses. This causes
issues if the page is loaded from some other website even though
the <cross-site-constraint> specified to be "*"