RFC 7232: Directory Service responds with 200 OK when If-Modified-Since matches Last-Modified from previous response · Issue #383 · kaazing/gateway · GitHub
Issue was caused by a difference of less than one second between the value
of the header  If-Modified-Since and the actual date of the modification
of the file.
Additionally the header Last-Modified was not sent when sending the status
304.

I have also took the liberty of closing certain FileInputStream which were
opened but never closed.