Http Connector authenticating to 401 and Login Module expiring state
The login modules will have an entry in their Login Map with the key "ExpiringState". The value of this entry will be the new public API of ExpiringState class. (Note: Java Doc needs to be public). This ExpiringState class is shared across all Login Modules and GW members in a cluster and can be used to share information across the login modules. (Like one time nonces, or multifactor credentials)

2) The http connector in the gateway can now respond to http challenges on 401 status codes. It will attempt to provide the credentials via a call to the Java Authenticator. The Authenticator is set on the JVM so we don't provide a GW specific way to configuring it and pointing it to related oracle docs on this should be sufficient. (For example Authenticator.setDefault). The http connector will only attempt to authenticate on the HTTP if a new connect-options property is set, Specifically this is "http.max.authentication.attempts" with an integer value. This specifies how many times an http connector will attempt to authenticate when it is challenged. By default this value is 0.