Fix the diagnostic messages from ExceptionMonitor for unexpected exception to include session details (#836)
Added a Kaazing version of ExceptionMonitor to mina.netty which adds a session parameter to method exceptionCaught. Set the (static) exception monitor in TransportFactory to an implementation that will include session details (including client identity) in the log message. Altered all code which calls the ExceptionMonitor to use the new version and pass in the session.